rabitMq/requirements.txt
pika
python-ulid[pydantic]
torch
easyocr
torchvision
supervision==0.18.0
openai==1.3.5
transformers
ultralytics==8.1.24
azure-identity
numpy
opencv-python
opencv-python-headless
gradio
dill
accelerate
timm
einops==0.8.0
paddlepaddle
paddleocr
uvicorn
fastapi
pydantic
pprintpp
requests 
Pillow 
python-dotenv
prisma
asyncio
langchain
langchain_ollama
redis
